requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
boto3==1.34.88
botocore==1.34.88
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.110.2
fastapi-storages==0.3.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
MarkupSafe==2.1.5
numpy==1.26.4
passlib==1.7.4
peewee==3.17.3
Pillow==9.5.0
psycopg2-binary
pydantic==1.9.2
pydantic_core==2.18.1
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
s3transfer==0.10.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
sqladmin==0.16.1
SQLAlchemy==1.4.40
starlette==0.37.2
threadpoolctl==3.4.0
typing_extensions==4.11.0
urllib3==1.26.18
uvicorn==0.18.2
WTForms==3.1.2
